nonneg input was changed to scanf'ing int

+++ lab_08_01/inc/input.h
@@ -9,10 +9,10 @@
 #define MAX_SIZE -1LLU
 
 int finput_size(FILE *const f, size_t *const size);
-int finput_nonneg(FILE *const f, size_t *const numb);
+int finput_nonneg(FILE *const f, int *const numb);
 int finput_matrix(FILE *const f, my_matrix_t matrix);
 int input_matrix(my_matrix_t matrix);
-int input_nonneg(size_t *const numb);
+int input_nonneg(int *const numb);
 int input_size(size_t *const size);
 void foutput_matrix(FILE *const f, my_matrix_t matrix);
 int input_power(int *const power);
